+++ src/Control/Actor/Language.hs
@@ -4,14 +4,26 @@
 module Control.Actor.Language where
 
 import           Universum
+import           Data.This
 import           Control.Monad.Free
 import           Control.Actor.Message
+import           Control.Concurrent.STM.TChan
+import           Control.Concurrent hiding (MVar, putMVar, takeMVar, newMVar)
+
+data Actor = Actor (TChan ActorMessage) ThreadId
 
 data ActorF next where
-    Math :: MessageType -> (ActorMessage -> IO ()) -> (() -> next)->  ActorF next
+    Math :: MessageType -> (ActorMessage -> IO ()) -> (() -> next) -> ActorF next
+    This :: (Actor -> next) -> ActorF next
     deriving (Functor)
 
 type ActorL = Free ActorF
+
+
+instance This ActorL Actor where
+    -- | Return link of current FSM.
+    this = liftF $ This id
+
 -- | Add handler for message wich known type.
 math :: Typeable a => (a -> IO ()) -> ActorL ()
 math f = liftF $ Math (fromActionToMessageType f) (f . fromActorMessageUnsafe) id