+++ ChangeLog.md
@@ -1,3 +1,12 @@
 # Changelog for Actor
 
+0.1 -> 0.2
+
+    [*] setFinisState => addFinalState
+    [+] add function “this” to actor language.
+
+    TODO:
+
+    notify -> emit
+
 ## Unreleased changes